+++ Cmdline/Action/Replace.cs
@@ -155,7 +155,8 @@ namespace CKAN.CmdLine
                 // TODO: These instances all need to go.
                 try
                 {
-                    ModuleInstaller.GetInstance(ksp, manager.Cache, User).Replace(to_replace, replace_ops, new NetAsyncModulesDownloader(User, manager.Cache));
+                    HashSet<string> possibleConfigOnlyDirs = null;
+                    ModuleInstaller.GetInstance(ksp, manager.Cache, User).Replace(to_replace, replace_ops, new NetAsyncModulesDownloader(User, manager.Cache), ref possibleConfigOnlyDirs);
                     User.RaiseMessage("\r\nDone!\r\n");
                 }
                 catch (DependencyNotSatisfiedKraken ex)