+++ GUI/MainDialogs.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Concurrent;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -45,46 +44,6 @@ namespace CKAN
         public int SelectionDialog(string message, params object[] args)
         {
             return selectionDialog.ShowSelectionDialog(message, args);
-        }
-
-        // Ugly Hack. Possible fix is to alter the relationship provider so we can use a loop
-        // over reason for to find a user requested mod. Or, you know, pass in a handler to it.
-        private readonly ConcurrentStack<GUIMod> last_mod_to_have_install_toggled = new ConcurrentStack<GUIMod>();
-
-        private async Task<CkanModule> TooManyModsProvideCore(TooManyModsProvideKraken kraken)
-        {
-            TaskCompletionSource<CkanModule> task = new TaskCompletionSource<CkanModule>();
-            Util.Invoke(this, () =>
-            {
-                UpdateProvidedModsDialog(kraken, task);
-                tabController.ShowTab("ChooseProvidedModsTabPage", 3);
-                tabController.SetTabLock(true);
-            });
-            return await task.Task;
-        }
-
-        public async Task<CkanModule> TooManyModsProvide(TooManyModsProvideKraken kraken)
-        {
-            // We want LMtHIT to be the last user selection. If we alter this handling a too many provides
-            // it needs to be reset so a potential second too many provides doesn't use the wrong mod.
-            GUIMod mod;
-
-            var module = await TooManyModsProvideCore(kraken);
-
-            if (module == null
-                    && last_mod_to_have_install_toggled.TryPeek(out mod))
-            {
-                MarkModForInstall(mod.Identifier, true);
-            }
-            Util.Invoke(this, () =>
-            {
-                tabController.SetTabLock(false);
-                tabController.HideTab("ChooseProvidedModsTabPage");
-                tabController.ShowTab("ManageModsTabPage");
-            });
-
-            last_mod_to_have_install_toggled.TryPop(out mod);
-            return module;
         }
     }
 }