+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - [Multiple] Custom mod labels, favorites, hiding (#2936 by: HebaruSan; reviewed: DasSkelett)
 - [GUI] Add "supports" relationships to recommendations screen (#2960 by: HebaruSan; reviewed: DasSkelett)
+- [Multiple] Prompt user to delete non-empty folders after uninstall (#2962 by: HebaruSan; reviewed: DasSkelett)
 
 ### Bugfixes